Rename 'install::compile' module to 'install::unix'
