+++ src/commands/install.rs
@@ -3,19 +3,15 @@ use semver::{Version, VersionReq};
 
 use crate::{commands, config::Cfg, settings::Settings, Result};
 
-mod compile;
 mod download;
 mod pip;
+mod unix;
 mod windows;
 
 use crate::commands::install::{
-    compile::{compile_source, extract_source},
     download::{download_source, find_all_python_versions},
     pip::install_extra_pip_packages,
 };
-
-#[cfg(target_os = "windows")]
-use crate::commands::install::windows::unattended_windows_install;
 
 pub fn run(
     from_version: Option<String>,
@@ -97,12 +93,12 @@ fn install_package(
 ) -> Result<()> {
     #[cfg(not(target_os = "windows"))]
     {
-        extract_source(&version_to_install)?;
-        compile_source(&version_to_install, install_extra_packages)?;
+        unix::extract_source(&version_to_install)?;
+        unix::compile_source(&version_to_install, install_extra_packages)?;
     }
     #[cfg(target_os = "windows")]
     {
-        unattended_windows_install(&version_to_install, install_extra_packages)?;
+        windows::unattended_windows_install(&version_to_install, install_extra_packages)?;
     }
 
     Ok(())
