+++ src/commands/install/unix.rs
